добавляет модули setup и data подключает точку входа
